+++ CHANGELOG.md
@@ -27,7 +27,9 @@
 
 ** 1.5.1 ** 
 
+* Added order lines for Klarna Pay Later payment method
+* Added order lines for AfterPay payment method
+* Fixed incorrect amount value of order
 * Klarna-Pay-Later: Remove fields gender and birthday from checkout form and customer object
 * Added the ability for AfterPay to be available in the selected countries.
 * Added the AfterPay localization for Netherlands, German and French language.
-